fix(slack): tighten tunnel install secret handling
